+++ des/src/simulation.cpp
@@ -5,9 +5,9 @@
 #include <vector>
 #include <iostream>
 
-Simulation::Simulation(std::vector<State*> states, std::vector<float> times): states(states), time(0), event_queue() {
+Simulation::Simulation(std::vector<State*> states, std::vector<double> times): states(states), time(0), event_queue() {
     int id;
-    std::vector<float>::iterator it;
+    std::vector<double>::iterator it;
     int first_state = 0; // assumption that everyone enters at state 0
 
     for (id=0, it = times.begin(); it != times.end(); ++it, ++id) {
@@ -32,7 +32,7 @@ void Simulation::run() {
     }
 }
 
-float Simulation::get_next_event_time() {
+double Simulation::get_next_event_time() {
     return event_queue.top()->time;
 }
 
@@ -46,11 +46,11 @@ State* Simulation::get_state(int index) {
     return states[index];
 }
 
-void Simulation::add_history(std::tuple<int, int, float> curr_state) {
+void Simulation::add_history(std::tuple<int, int, double> curr_state) {
     history.push_back(curr_state);
 }
 
-std::vector<std::tuple<int, int, float> > Simulation::get_history() {
+std::vector<std::tuple<int, int, double> > Simulation::get_history() {
     return history;
 }
