+++ des/src/state.h
@@ -5,14 +5,14 @@
 #include <utility>
 #include "transitions.h"
 
-const float MINIMUM_EVENT_TIME = 1.0f;
+const double MINIMUM_EVENT_TIME = 1.0;
 
 class State {
     public:
         const int num;
         State(int num);
         ~State(void);
-        std::pair<int, float> get_next_transition(int);
+        std::pair<int, double> get_next_transition(int);
         void add_transition(Transition*);
         bool is_transient() const;
